can eth, finish

+++ api/domain/database.define.js
@@ -107,7 +107,7 @@ model.DomainEthListener = sequelize.define("t_listener_eth",{
         field:"tx_to"
     },
     txValue:{
-        type:Sequelize.BIGINT,
+        type:Sequelize.DECIMAL(40,8),
         field:"tx_value"
     },
     txInput:{
@@ -167,7 +167,7 @@ model.DomainBtcListener = sequelize.define("t_listener_btc",{
         field:"tx_to"
     },
     txValue:{
-        type:Sequelize.BIGINT,
+        type:Sequelize.DECIMAL(40,8),
         field:"tx_value"
     },
     txInput:{
